requirements.txt
pip~=24.0
fastapi==0.110.0
uvicorn==0.28.0
python-dotenv==1.0.1
